rename util files to kebab-case
