+++ packages/hash/playwright/tests/page-readonly-mode.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from "@playwright/test";
-import { loginUsingUi } from "./utils/loginUsingUi";
-import { resetDb } from "./utils/resetDb";
+import { loginUsingUi } from "./utils/login-using-ui";
+import { resetDb } from "./utils/reset-db";
 
 test.beforeEach(async () => {
   await resetDb();
